Mechanical move of Check functions to tx_check.go file.
